Switch to use video_nr 10 and remove removed CLI options (#217)
It is quite likely for people to already have video2 and/or video3 occupied so
better to take much higher default number.  This would lead to default
conflicts. Also having a distinctive video10 makes is easier to identify this
camera.

And then it might also be more likely to have video0 available for the default
video camera, thus better to use that video2 or any other.